LE_Kotlin: created new package "geometry" and moved relevant classes there
