+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/Coordinates3D.kt
@@ -1,4 +1,7 @@
-package elements
+package geometry
+
+import elements.*
+import elements.Border
 
 // 3d as simple as 2d :)
 data class Coordinates3D(
